+++ VirtualList/useSpotlight.js
@@ -6,9 +6,9 @@ const useSpotlightConfig = (props, instances) => {
 	// Hooks
 
 	useEffect(() => {
-		const {spottable: {current: {lastFocusedIndex}}} = instances;
+		function lastFocusedPersist () {
+			const {spottable: {current: {lastFocusedIndex}}} = instances;
 
-		const lastFocusedPersist = () => {
 			if (lastFocusedIndex != null) {
 				return {
 					container: false,
@@ -16,21 +16,38 @@ const useSpotlightConfig = (props, instances) => {
 					key: lastFocusedIndex
 				};
 			}
-		};
+		}
+
+		function lastFocusedRestore ({key}, all) {
+			const placeholder = all.find(el => 'vlPlaceholder' in el.dataset);
+
+			if (placeholder) {
+				placeholder.dataset.index = key;
+
+				return placeholder;
+			}
+
+			return all.reduce((focused, node) => {
+				return focused || Number(node.dataset.index) === key && node;
+			}, null);
+		}
 
 		function configureSpotlight () {
 			const {spacing, spotlightId} = props;
 
 			Spotlight.set(spotlightId, {
+				enterTo: 'last-focused',
 				/*
 				 * Returns the data-index as the key for last focused
 				 */
 				lastFocusedPersist,
+
 				/*
 				 * Restores the data-index into the placeholder if its the only element. Tries to find a
 				 * matching child otherwise.
 				 */
 				lastFocusedRestore,
+
 				/*
 				 * Directs spotlight focus to favor straight elements that are within range of `spacing`
 				 * over oblique elements, like scroll buttons.
@@ -41,26 +58,6 @@ const useSpotlightConfig = (props, instances) => {
 
 		configureSpotlight();
 	}, [props, instances]);
-
-	// Functions
-
-	/*
-	 * Restores the data-index into the placeholder if it exists. Tries to find a matching child
-	 * otherwise.
-	 */
-	function lastFocusedRestore ({key}, all) {
-		const placeholder = all.find(el => 'vlPlaceholder' in el.dataset);
-
-		if (placeholder) {
-			placeholder.dataset.index = key;
-
-			return placeholder;
-		}
-
-		return all.reduce((focused, node) => {
-			return focused || Number(node.dataset.index) === key && node;
-		}, null);
-	}
 };
 
 const getNumberValue = (index) => index | 0;