+++ tests/ui/specs/Scroller/ScrollerPage.js
@@ -1,6 +1,11 @@
 'use strict';
 const {element, Page} = require('@enact/ui-test-utils/utils');
 
+const focusableBodySelector = '.Scroller_Scroller_focusableBody';
+const scrollbarSelector = '.useScroll_ScrollbarTrack_scrollbarTrack';
+const scrollVirticalThumbSelector = '.useScroll_ScrollbarTrack_vertical  .useScroll_ScrollbarTrack_thumb';
+const scrollHorizontalThumbSelector = '.useScroll_useScroll_horizontalScrollbar .useScroll_ScrollbarTrack_thumb';
+
 class ScrollerPage extends Page {
 
 	constructor () {
@@ -12,10 +17,48 @@ class ScrollerPage extends Page {
 		super.open(`Scroller${layout}-View`, urlExtra);
 	}
 
-	get button1 () { return element('#Page_1_Button', browser); }
-	get button2 () { return element('#Page_2_Button', browser); }
-	get button3 () { return element('#Page_3_Button', browser); }
+	// button api
+	get button1 () {
+		return element('#Page_1_Button', browser);
+	}
+	get button2 () {
+		return element('#Page_2_Button', browser);
+	}
+	get button3 () {
+		return element('#Page_3_Button', browser);
+	}
+	get buttonTop () {
+		return element('#top', browser);
+	}
+	get buttonHideScrollbar () {
+		return element('#hideScrollbar', browser);
+	}
 
+	// dropdown api
+	get dropdownFocusableScrollbar () {
+		return element('#focusableScrollbarKnobs', browser);
+	}
+
+	// scrollable api
+	get focusableBody () {
+		return $(`${focusableBodySelector}`);
+	}
+	get verticalScrollThumb () {
+		return $(`${scrollVirticalThumbSelector}`);
+	}
+	get horizontalScrollThumb () {
+		return $(`${scrollHorizontalThumbSelector}`);
+	}
+	/* global document */
+	getScrollThumbPosition () {
+		return browser.execute(function (_scrollbarSelector) {
+			const scrollbar = document.querySelectorAll(_scrollbarSelector);
+			return {
+				vertical: scrollbar[0].style.getPropertyValue('--scrollbar-thumb-progress-ratio'),
+				horizontal: scrollbar[1].style.getPropertyValue('--scrollbar-thumb-progress-ratio')
+			};
+		}, scrollbarSelector);
+	}
 }
 
 module.exports = new ScrollerPage();