+++ tests/ui/specs/VirtualList/VirtualList/VirtualList-specs.js
@@ -1,5 +1,5 @@
 const Page = require('./VirtualListPage'),
-	{expectFocusedItem, expectNoFocusedItem, waitForScrollStartStop, waitUntilFocused} = require('../VirtualList-utils');
+	{expectFocusedItem, expectNoFocusedItem, waitUntilFocused} = require('../VirtualList-utils');
 
 describe('VirtualList', function () {
 
@@ -19,55 +19,18 @@ describe('VirtualList', function () {
 			expectFocusedItem(0);
 		});
 
-		it.skip('should focus and Scroll with Up/Down and 5-way [GT-28491]', function () {
-			Page.buttonLeft.moveTo();
-			Page.spotlightRight(); // is on 'Item 000'
-			// Step 3. 5-way Spot the second item 'Item 001'.
-			Page.spotlightDown();
-			// Verify Step 3: Spotlight displays on the second item 'item 001'.
-			expectFocusedItem(1, 'step 3 focus');
-			// Step 4. Press Channel Down.
-			Page.pageDown();
-			waitForScrollStartStop();
-			// Verify Step 4: Spotlight is on the *Item* closest to the previously focused Item's location.
-			waitUntilFocused(7, 'step 4 focus'); // this works in headless + tv  - must comment to run in debug
-			// Step 5. 5-way Down several times to the last visible item on the current viewport.
-			Page.fiveWayToItem(17);
-			// Verify Step 5: Spotlight is on the last visible item. *** it is not
-			Page.delay(100);
-			expectFocusedItem(17, 'step 5 focus');
-			// Step 6. Press Channel Down.
-			Page.pageDown();
-			waitForScrollStartStop();
-			// Verify Step 6: Spotlight is on the *Item* closest to the previously focused Item's location  ?
-			waitUntilFocused(23, 'step 6 focus');
-			// Step 7. Press Channel Up.
-			Page.pageUp();
-			waitForScrollStartStop();
-			// Verify Step 7: Spotlight is on the *Item* closest to the previously focused Item's location.
-			waitUntilFocused(17, 'step 7 focus');
-			// Step 8. 5-way Up several times to the first visible item on the current viewport.
-			Page.fiveWayToItem(7);
-			// Verify Step 8: Spotlight is on the first visible item.
-			expectFocusedItem(7, 'step 8 focus');
-			// Step 9. Press Channel Up.
-			Page.pageUp();
-			waitForScrollStartStop();
-			// Verify Step 9: Spotlight is on the *Item* closest to the previously focused Item's location.
-			waitUntilFocused(1, 'step 9 focus');
-			// Step 10. Wave the pointer. Step 11. Hover on an item.
-			Page.item(3).moveTo();
-			// Verify Step 10, Step 11: Spotlight is on 'Item 003'
-			expectFocusedItem(3, 'step 11 focus');
-			// Step 12. Press Channel Down.
-			Page.pageDown();
-			waitForScrollStartStop();
-			// Verify Step 12: 1. Spotlight hides. 2. The list Scrolls Up by page with animation. 3. The list stops scrolling. 4. Spotlight still hides (for a few seconds).
-			expectNoFocusedItem();  // Check that Spotlight hides only.
-		});
-
 		it('should not scroll when leaving list with 5-way up/down [GT-28473]', function () {
-			// Step 3. Set dataSize to 100. Step 4: change to 5-way mode
+			// Step 3 is 'Set dataSize to 100' but set dataSize to 20 for Speed up Test.
+			Page.inputfieldNumItems.moveTo();
+			Page.spotlightSelect();
+			Page.backSpace();
+			Page.backSpace();
+			Page.backSpace();
+			Page.numPad(2);
+			Page.numPad(0);
+			Page.backKey();
+			Page.spotlightDown();
+			// Step 4: change to 5-way mode
 			Page.buttonLeft.moveTo();
 			// Step 5: 5-way Spot the first item.
 			Page.spotlightRight();
@@ -81,12 +44,16 @@ describe('VirtualList', function () {
 			// Page.mouseWheel(40, Page.item(6));   currently not working as expected so using 5-way Down temporary
 			// Wheeling will not be implemented - see ENYO-6178
 			Page.spotlightDown();
-			Page.fiveWayToItem(99);
+			expectFocusedItem(0);
+			Page.pageDown();
+			waitUntilFocused(6, 'focus Item 6');
+			Page.pageDown();
+			waitUntilFocused(19, 'focus Item 19');
 			// Step 7: 2. Click the last item.
 			Page.spotlightSelect();
 			// Verify Step 7: Spotlight is on the last item.
 			Page.delay(1000);
-			expectFocusedItem(99, 'step 7 focus');
+			expectFocusedItem(19, 'step 7 focus');
 			// Step 8: 5-way Down
 			Page.spotlightDown();
 			Page.spotlightDown(); // 1 extra 5-way down to check Spotlight does not pass buttonBottom when wrap is off.
@@ -96,12 +63,6 @@ describe('VirtualList', function () {
 			expect(Page.buttonBottom.isFocused(), 'step 8 focus').to.be.true();
 		});
 
-		//	TODO: Need to api for Scrollbar and List size checking in sandstone.
-		it.skip('should have same height list and scrollbar [GT-28930]', function () {
-			// Verify: The scrollbar size fit to the size of the list.
-			expect(Page.listSize.height).to.equal(Page.scrollBarSize.height);
-		});
-
 		it('should position Scrollbar Track on right side in LTR [GT-28562]', function () {
 			let ListwidthSize = Page.getScrollOffsetLeft() + Page.getScrollbarWidth();
 			// Verify Step 2.2: The Scrollbar track displays shortly right aligned.
@@ -109,11 +70,20 @@ describe('VirtualList', function () {
 		});
 
 		it('should position Scroll thumb on top/bottom when reaching to the edge with 5-way and Channel Down [GT-28564]', function () {
-			// Test (Jira) calls for 30 items only. Test uses default of 100 items.
+			// Step 3. Knobs > VirtualList > dataSize > 30
+			Page.inputfieldNumItems.moveTo();
+			Page.spotlightSelect();
+			Page.backSpace();
+			Page.backSpace();
+			Page.backSpace();
+			Page.numPad(3);
+			Page.numPad(0);
+			Page.backKey();
+			Page.spotlightDown();
 			// Step 4. Move focus to the first item ('Item 00').
-			// Verify Step 4: 1. Spotlight displays on the first item.
 			Page.buttonLeft.moveTo();
 			Page.spotlightRight();
+			// Verify Step 4: 1. Spotlight displays on the first item.
 			expectFocusedItem(0, 'focus Item 0');
 			// Verify Step 5: Scroll thumb's position appears shortly at the top of the Scrollbar track.
 			expect(Page.getScrollThumbPosition(), 'Up').to.be.equal('0');
@@ -128,23 +98,25 @@ describe('VirtualList', function () {
 			expectNoFocusedItem();
 			waitUntilFocused(12, 'focus Item 12');
 			// Step 8. 5-way Down several times to scroll down the list.
-			Page.fiveWayToItem(30);
-			expectFocusedItem(30, 'focus Item 30');
+			Page.fiveWayToItem(20);
 			// Step 9. 5-way Spot the last item.
-			Page.fiveWayToItem(99);
+			Page.fiveWayToItem(29);
 			// Verify Step 9: 1. Spotlight displays on the last item.
+			waitUntilFocused(29, 'focus last Item');
 			Page.delay(1000);
-			expectFocusedItem(99, 'focus Item 99');
 			// Verify Step 10: Scroll thumb's position appears shortly at the bottom of the Scrollbar track.
-			Page.delay(2000);
 			expect(Page.getScrollThumbPosition(), 'Down').to.be.equal('1');
 			// Step 11: 5-way Spot the first item.
-			Page.fiveWayToItem(0);
+			Page.pageUp();
+			waitUntilFocused(23, 'focus Item 23');
+			Page.pageUp();
+			waitUntilFocused(17, 'focus Item 17');
+			Page.pageUp();
+			waitUntilFocused(11, 'focus Item 11');
+			Page.pageUp();
 			// Verify Step 11: Spotlight displays on the first item.
-			Page.delay(2000);
-			expectFocusedItem(0, 'focus Item 0');
+			waitUntilFocused(0, 'focus Item 0');
 			// Verify Step 12: Scroll thumb's position appears shortly at the top of the Scrollbar track.
-			Page.delay(1000);
 			expect(Page.getScrollThumbPosition(), 'Up').to.be.equal('0');
 		});